feat(接口测试): 接口测试获取环境详情信息
